Added Firebase utils class.
It contains functions for getting and setting data from Firebase database.
The parameters given to these functions should be a List containing the path to the requested resource.
For example, if you want to modify user1's id, the function call should look like
	FirebaseUtils.setValue(Arrays.asList("User", "user1", "id"), newValue);

Also modified the login function to use the FirebaseUtils methods.

Signed-off-by: Cristi Iacob <cristiiacob98@gmail.com>